futher optimization

+++ src/utils.jl
@@ -286,8 +286,8 @@ end
 
 function chempotk_simple(psi, sim)
     @unpack ksquared, dV, V0, Vol, g, equation = sim
-    mu = 1 / Vol * sum(1 / 2 * ksquared .* abs2.(psi))
-    tmp = xspace(psi, sim)
+    mu::Float64 = 1 / Vol * sum(1 / 2 * ksquared .* abs2.(psi))
+    tmp::Float64 = xspace(psi, sim)
     mu += dV * sum((V0 + g * abs2.(tmp)) .* abs2.(tmp))
     mu += 1 # add one transverse energy unit (1D-GPE case)
     return mu
